feat: 사진 좌표 추적
- 사진 좌표 추적 구현 완료
- 다른 위젯을 메인 위젯에 불러와 해결
- 사이즈 맞춰준 후 출력해야 정상출력됨
