+++ View.py
@@ -1,5 +1,7 @@
 from PyQt5 import uic
 from PyQt5.QtWidgets import *
+from PyQt5.QtCore import *
+from PyQt5.QtGui import *
 import sys
 
 from ViewModel.Canvas import Canvas
@@ -19,8 +21,6 @@ class HandAnnot(QMainWindow, main_form_class):
 
         canvas_widget = [
                             self.action_Open,
-                            self.label_Canvas,
-                            self.scrollArea_Canvas,
                             self.menu_Edit,
                             self.menu_Zoom,
                             self.action_Save,
@@ -31,21 +31,7 @@ class HandAnnot(QMainWindow, main_form_class):
 
         self.canvas_viewmodel = Canvas(canvas_widget, self.model)
 
-    def mouseMoveEvent(self, event):
-        x_pos = event.x() - 20
-        y_pos = event.y() - 50
-        self.model.setCurPos([x_pos, y_pos])
-        text = '{x_pos}, {y_pos}'.format(x_pos=x_pos, y_pos=y_pos)
-        self.statusBar.showMessage(text)
-    
-    def mouseReleaseEvent(self, event):
-        if self.hasMouseTracking():
-            self.setMouseTracking(False)
-        else:
-            past_x_pos = event.x() - 20
-            past_y_pos = event.y() - 50
-            self.model.setPrePos([past_x_pos, past_y_pos])
-            self.setMouseTracking(True)
+        self.scrollArea_Canvas.setWidget(self.canvas_viewmodel)
 
 if __name__=='__main__':
     app = QApplication(sys.argv)
